added functionality to follow politicians and refactored favorite companies
