+++ src/client/features/companies/CompanyCard.jsx
@@ -1,55 +1,15 @@
 import React, { useEffect } from "react";
 import { useSelector } from "react-redux";
-import {
-  selectToken,
-  selectUserId,
-  useGetAccountQuery,
-} from "../auth/authSlice";
+import { selectToken } from "../auth/authSlice";
 import { useParams, Link } from "react-router-dom";
 import { useGetCompanyByIdQuery } from "./companySlice";
-import {
-  useAddFavoriteCompanyMutation,
-  useFetchFavoriteCompaniesQuery,
-  useRemoveFavoriteCompanyMutation,
-} from "../Account/favorites/favSlice";
+import { useFavorites } from "../Account/favorites/favUtility";
 
 export default function CompanyCard({ company }) {
   const { id } = useParams();
-  const userId = useSelector(selectUserId);
   const token = useSelector(selectToken); 
-  const [addFavoriteCompany] = useAddFavoriteCompanyMutation();
-  const [removeFavoriteCompany] = useRemoveFavoriteCompanyMutation();
+  const { handleAddFavorite, favoriteCompanies } = useFavorites();
 
-  
-  const { data: user } = token ? useGetAccountQuery(userId) : { data: null };
-  const { data: favoriteCompanies, refetch: refetchFavorites } = token ? useFetchFavoriteCompaniesQuery(userId) : { data: null, refetch: null };
-
-
-  const handleAddFavorite = async () => {
-    try {
-      if (company && user && favoriteCompanies) {
-        const userId = user.id;
-        const companyId = company.id;
-
-        const isCompanyFavorite = favoriteCompanies.some(
-          (favorite) => favorite.companyId === company.id
-        );
-
-        if (isCompanyFavorite) {
-          // If the company is already a favorite, remove it
-          await removeFavoriteCompany({ userId, companyId });
-        } else {
-          // If the company is not a favorite, add it
-          await addFavoriteCompany({ userId, companyId });
-        }
-
-        // Refetch favorite companies to update the data
-        refetchFavorites();
-      }
-    } catch (error) {
-      console.error("Error updating favorite status:", error);
-    }
-  };
 
   if (id !== undefined) {
     const { data: companyData, isLoading } = useGetCompanyByIdQuery(id);
@@ -77,11 +37,11 @@ export default function CompanyCard({ company }) {
             <p>Headquarter: {companyData.hq}</p>
             <p>Founded at year: {companyData.founded}</p>
             {token && (
-          <button className="favButton" onClick={handleAddFavorite}>
-            {favoriteCompanies && favoriteCompanies.some((favorite) => favorite.companyId === companyData.id)
-              ? "Remove from Favorites"
-              : "Add to Favorites"}
-          </button>
+          <button className="favButton" onClick={() => handleAddFavorite(companyData)}>
+          {favoriteCompanies && favoriteCompanies.some((favorite) => favorite.companyId === companyData.id)
+            ? "Remove from Favorites"
+            : "Add to Favorites"}
+        </button>
           )}
           </div>
         )}
@@ -96,11 +56,11 @@ export default function CompanyCard({ company }) {
           <p> {company.security}</p>
           <Link className="companyCard-Link" to={`/companies/${company.id}`}>More Info</Link>
           {token && (
-          <button className="favButton" onClick={handleAddFavorite}>
-            {favoriteCompanies && favoriteCompanies.some((favorite) => favorite.companyId === company.id)
-              ? "Remove from Favorites"
-              : "Add to Favorites"}
-          </button>
+          <button className="favButton" onClick={() => handleAddFavorite(company)}>
+          {favoriteCompanies && favoriteCompanies.some((favorite) => favorite.companyId === company.id)
+            ? "Remove from Favorites"
+            : "Add to Favorites"}
+        </button>
           )}
         </div>
       </>
